requirements.txt
streamlit==1.33.0
mysql-connector-python==8.3.0
pandas==2.2.2
Pillow==10.3.0


#to instal --> pip install -r requirements.txt
# with best , Abdalkarim Dwikat